+++ src/Sitecore.Media.AzureBlobStorage.sln
@@ -1,0 +1,27 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sitecore.Media.AzureBlobStorage", "Sitecore.Media.AzureBlobStorage\Sitecore.Media.AzureBlobStorage.csproj", "{B5761D5E-0CD9-44E4-85AF-DFA8D53A7836}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "content", "content", "{69629A3F-A2A1-444E-99F4-405CD5C31B1C}"
+	ProjectSection(SolutionItems) = preProject
+		..\README.md = ..\README.md
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B5761D5E-0CD9-44E4-85AF-DFA8D53A7836}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5761D5E-0CD9-44E4-85AF-DFA8D53A7836}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5761D5E-0CD9-44E4-85AF-DFA8D53A7836}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5761D5E-0CD9-44E4-85AF-DFA8D53A7836}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
