+++ src/Sitecore.Media.AzureBlobStorage/App_Config/Include/Sitecore.AzureExtensions.config
