+++ src/Sitecore.Media.AzureBlobStorage/packages.config
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="HtmlAgilityPack" version="1.4.6" targetFramework="net45" />
+  <package id="Lucene.Net" version="3.0.3" targetFramework="net45" />
+  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net45" />
+  <package id="Microsoft.Data.Edm" version="5.6.4" targetFramework="net45" />
+  <package id="Microsoft.Data.OData" version="5.6.4" targetFramework="net45" />
+  <package id="Microsoft.Data.Services.Client" version="5.6.4" targetFramework="net45" />
+  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="1.8.0.0" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net45" />
+  <package id="SharpZipLib" version="0.86.0" targetFramework="net45" />
+  <package id="Sitecore.Kernel" version="8.0.150427" targetFramework="net45" developmentDependency="true" />
+  <package id="Sitecore.Logging" version="8.0.150427" targetFramework="net45" developmentDependency="true" />
+  <package id="Sitecore.NVelocity" version="8.0.150427" targetFramework="net45" developmentDependency="true" />
+  <package id="Sitecore.Zip" version="8.0.150427" targetFramework="net45" developmentDependency="true" />
+  <package id="System.Spatial" version="5.6.4" targetFramework="net45" />
+  <package id="WindowsAzure.Storage" version="7.2.0" targetFramework="net45" />
+</packages>
